Move fluxUtil to rx package
