+++ app/src/main/kotlin/com/adriangl/pokeapi_mvvm/pokemonlist/PokemonActions.kt
@@ -10,6 +10,3 @@ class GetPokemonDetailsListAction : BaseAction()
 data class PokemonDetailsListLoadedAction(val pokemonList: List<Pokemon>?, val task: Task)
 
 data class FilterPokemonListAction(val query: String) : BaseAction()
-
-data class PokemonListFilteredAction(val query: String?, val list: List<Pokemon>?, val task: Task)
-
